Don't read stats for fulltext or spatial indexes
This silences a Valgrind warning about conditional jump depending on
uninitialized values. We do not maintain stats for fulltext or spatial
indexes and thus dict_index_t::stat_n_leaf_pages for such an index is
left uninitialized when initializing the rest of the stats of the table
and its indexes.

Reviewed by: Marko Makela <marko.makela@oracle.com>