Bug#21464737: MTS SUBMODE ON WORKERS DOES NOT RESPECT THE LOCAL CHANNEL VALUE.
Since the introduction of the channel service interface, the developer
can now set the MTS submode per channel being the value stored in the
channel Relay_Log_Info object.

This value is however ignored in the workers (rpl_rli_pdb.cc). The
workers still read and use the global flag set for all slave
channels. This can lead to error as the coordinator can be of one
type and its workers of another.
This is now corrected as the workers read the value on the rli object.