Bug #20229614 : OR CONDITIONS ON MULTI-COLUMN INDEX MAY NOT                 USE ALL INDEX COLUMNS TO FILTER ROWS
ISSUE:
------
While optimizing a range scan for the OR-operator, key_or
incorrectly assumes an out-of-memory situation. This in turn
results in an incomplete condition being considered for the
query plan and gives an incorrect row estimate. Hence the
over-estimation of the number of qualifying rows.


SOLUTION:
---------
In key_or(), a key condition is cloned if it's already in
use in some other condition. In this case NULL should be
returned only if allocation fails.

The check of key_count > 0 was originally added as part of
Bug #4157. This was done to check the key_count after the
swap of keys. Here changing the "||" to "&&" ensures that
when the new key_count is greater than 0, there will be
an attempt to clone the tree.

This also exposes a different issue. When two SEL_ARG
predicates are "Or-ed" together, i.e. combined into a
common OR expression, the case when either predicate was of
the type ALWAYS, i.e. unable to reject any rows, was not
handled. When key_or tries to clone a SEL_ARG object of the
type ALWAYS, it results in a failed assertion.

This is fixed by treating ALWAYS SEL_ARG's similar to NULL
SEL_ARG's. If either SEL_ARG type is ALWAYS, return it and
release the reference to other one. The logic is similar to
the inference TRUE or P => P.