Bug#21338147  INNODB: FAILING ASSERTION: !TRX_IS_AUTOCOMMIT_NON_LOCKING(TRX) IN TRX0TRX.CC
To a failed single statement transaction with a specified gtid on
auto-commit mode, we missed to release its owned gtid when
rolling back it if binlog is disabled. Which causes the
assertion failure.

To fix the problem, to a failed single statement transaction with a
specified gtid on auto-commit mode, we roll back its owned gtid if
it does not modify non-transational table or commit its owned gtid
if it has modified non-transactional table when rolling back it if
binlog is disabled, as we did when binlog is enabled. Then it does
not cause an assertion failure any more.