Patch for bug#21671754
NEXTEVENT() SAYS:'!(GCI_OPS && (GCI == GCI_OPS->M_GCI)' ON A TE_NUL EVENT

TE_NUL events are ignored, and intended as an internal placeholder
only. They should not be expected to be 'consistent' in the sense
that all relevant data for them are properly filled out. Hence. it
does not make sense to check that the GCIs are in synch between
'data' and the 'gci_ops-list' for these events.

Thus, move the GCI check to after TE_NUL are skipped.