Fix for bug#21673318
 	EXCEPTIONAL EVENTS WHICH ARE DROPPED, SHOULD NOT BE RETURNED

The event-API previously returned exceptional events
which was attached to event operations which had been dropped
in some circumstances.

However in other cases, when *all* event operations were dropped,
it returned only those which were polled before the drop,
while the unpolled were ignored.

Furthermore, even the polled ex-events were discarded if
a new event op was later created.

This was a consequence if the ex-events being associated
with an event operation when inserted into the event queue.

This fix defers the association between the ex-event and
event operation intil the event is consumed by nextEvent().
It will then find a non-dropped event operation to assiciate
it with. If all event operations were dropped, the ex-event
is ignored.