Bug#21530366 CRASH/ASSERTION, CORRUPTION WITH INDEXES + VIRTUAL COLUMNS, BLOB
This crash happens when InnoDB tries to update a secondary index for a
virtual field of type BLOB. When doing this update, InnoDB relies on
that both the old and new value for the BLOB is available. The crash
occurred due to the server provided a wrong value as the old BLOB
value. In the case of the failing query in the test case, the server
had over-written the old BLOB value with the new BLOB value.

Before the server updates a virtual BLOB field, it sets the
Field_blob::keep_old_value to true. In Field_blob::store_internal()
there is code that checks this flag and if it is set, it will not
over-write the current BLOB value. But in the case of the failing
query, the current BLOB value is over-written before we reach the code
that checks the Field_blob::keep_old_value flag. Both in field_conv()
and early in Field_blob::store_internal() we have code that can
over-write the current BLOB value in the case where the new BLOB value
is written to the same memory as the current BLOB value.

The fix for this problem is to replace the Field_blob::keep_old_value
flag with a new function that saves the current version of the blob
value. In update_generated_write_fields() we call this function
instead of setting the keep_old_value flag before calling
Item::save_in_field(). By doing this we ensure that the old value of
the blob is safely stored before Item::save_in_field() can do any
changes to it.