Implement reserved SendBuffer concept (bug#21664515)
This patch adds the notion of some part of the
TF SendBuffer pool being reserved.  This is
done by allowing allocations to use the last
n pages only if their try_alloc() call has the
reserved bool set to true.

The intention is that the loopback transporter
can set the reserved bool when allocating,
other transporters cannot, so the loopback
transporter has some region of the pool which
is protected from the other users.

Some complexity could be removed by using
e.g. a try_alloc(1) specific method -
try_alloc(n>1) does not appear to be used
currently.

Also note that the total and inuse bytes
refer to pages allocated rather than useful
bytes on those pages.  It's not clear that
these are used for anything currently, but if
they are, then the desired semantics should be
considered - should this include reserved space
or not?