Avoid processing DISCONNECT_REP twice (bug#21683144)
ClusterMgr assumes that reportConnected() /
reportDisconnected() will be serialised w.r.t.
each other, but this is not the case where a
node with performState==CONNECTING is
transitioned to DISCONNECTING.

In this case it is possible for it to then
transition to DISCONNECTED, and
reportDisconnected() to be called without a
preceding call to reportConnected().

This was causing assertion failures on a
pre-decrement assert on noOfConnectedNodes.
It may have had other bad effects with assertions
off.

The solution chosen here is to only process
DISCONNECT_REP if reportConnected() was processed
- e.g. if the connection succeeded.  If the
connection did not succeed then the ClusterMgr
has no real state to cleanup.