Bug#21681433 FORMATTING ERROR IN OS_FILE_WRITE_PAGE()
Problem:
========
The error messages in os_file_write_page() has a left over "%d" in the
message (possibly during the conversion from C-style (fprintf()) to
C++ Style (ib::errror() << "..")) and error messages show it.

Also, the use of '.' , in the same function, to separate 'offset' value
(file offset from the start where to read) and 'n' value (number of bytes
to read, starting from offset) is quite confusing.

Fix:
====
Removed the "%d" in the error message and used ', ' instead of '.' for
better understanding.

Reviewed-by: Kevin Lewis <kevin.lewis@oracle.com>
RB: 10051