BUG#21631197 - ASSERT 0 IN INNODB.DOUBLEWRITE RUNNING UNDER VALGRIND
This seems only affects the test cases of user table. I can easily break
the test case by adding some 'sleep N;' before the kill_mysqld.

The reason of the failure is that if there is some break between marking
a specified dirty page and the killing server, there could be a checkpoint,
which marks the checkpoint LSN to be a bigger one than the LSN generated
by dirtying the page. So during recovery, server always consider that
this dirty page should be correct and safe to read, no need to read it
from double write buffer. But in fact, we will modify it after kill_mysqld,
and this breaks the rule.

The solution is that we should make sure dirtying a specified page is
the last operation before kill_mysqld, and there should be no checkpoints.

Reviewed-by: Jimmy Yang <jimmg.yang@oracle.com>
RB: 10004