Fix a race between delete and copy-to-next
There is a bug in the lock free hash table used in WL#7170:

During copy to the next array tuples with val = DELETED are skipped and
not copied as an optimization because such tuples are treated as they do
not exist, so their existence is not necessary and they only take up
space.

But there is the following race between copy and concurrent delete:

copy: copy to the next array because val != DELETED (e.g val == 12345)
(now the next array contains the tuple with val == 12345)

del: change val from 12345 to DELETED

copy: try to CAS val from 12345 to GOTO_NEXT_ARRAY, but it fails because
      val is now DELETED. CAS will return the most recent val and it
      will be DELETED

copy: since CAS failed and we have the most recent value (DELETED), copy
      again to the next array, but val == DELETED and we skip this
      operation as the buggy optimization, nothing is done

copy: try to CAS val from DELETED to GOTO_NEXT_ARRAY, success.

Now the old array contains a tuple with val == DELETED, the next array
contains a tuple with val == 12345 and the effect of the delete
operation is lost.

WL#7170 InnoDB buffer estimates for tables and indexes