Patch for bug#21689380, EVENT-API, MEMORY LEAK OF GCI_OP(S)
Patch changes NdbEventBuffer::free_list() such that it
now handles being called for an empty list of EventBufData.
The gci_op(s) might still be no-empty though, which was also
handled by ::free_list() ... if called

Callees of ::free_list() is changed such that it is
now called unconditionally, and the explicit handling of
dangling gci_op(s), where it was handled at all, is removed.
(As it is also handled by free_list())

Also removed an obsolete call to clear_event_queue() from
NdbEventBuffer d'tor, as freeing of m_available_data was already
handled.