Bug#21263152 ST_SYMDIFFERENCE(POLYGON, POLYGON) MAY RETURN AN INVALID MULTIPOLYGON
Problem: Some queries with ST_SymDifference may return invalid
multipolygons in the result.

The reason is bug 11436 in Boost.Geometry.

Fix: Backport bugfix from BG. The symmetric difference operation for
two areal geometries A1 and A2 is now implemented as follows:

1. Compute the difference of A1 minus A2
2. Compute the difference of A2 minus A1
3. Return the union of the two differences computed in the two steps
   above.

The BG bugfix reveals a new problem in the Gis_wkb_vector copy
constructor that causes it to copy only the exterior ring of a
polygon, ignoring the inner rings.

Even if the m_geo_vect of the object being copied points to a vector,
the constructor only checks get_ptr(), which is NULL if the object has
been constructed by BG and reassemble() has not yet been called.

Fix: Make the copy constructor check for m_geo_vect in addition to
get_ptr().