Bug #19333852: RESULT DIFF IN QUERY HAVING DISTINCT WITH                GROUP BY
Issue:
-----
This problem happens under the following conditions:
1) Range optimization is attempted on an index with a
   string column as its first part.
2) The value used for comparision is fully truncated.

This makes comparisions with index incorrect. The range
optimizer incorrectly concludes that there are no values
in the index that satisfy the relevant comparision.

SOLUTION:
---------
We keep track of whether a string has been fully truncated
during range optimization. If yes, that string is not used
as part of the range optimization to avoid incorrect
filtering.