Bug#21652012 GIS FUNCTIONS RETURN ERROR WITH VALID INPUT WHEN USED IN SUBQUERY WITH NOT IN
Issue:
The simplify_multi_geometry() simplifies a GEOMETRY buffer by directly modifying
the byte string, however, this is not OK if the byte string is used again, as in
the test cases of this bug report. The reason is that the provider of the
GEOMETRY buffer assumes it intact(including its content&length).

Fix:
Make a copy of the GEOMETRY byte string before actually simplifying it, and
use this copy instead. The copy is only made if absolutely necessary.