Bug #18698556: UPDATE ORDER BY DOES A FILESORT IF UPDATING IN THE INDEX
Problem:
Filesort is being done for an update query having order by,
when the key being used for scanning is also updated.

Analysis:
Whenever a key that could be used for ordering is identified,
optimizer also checks if this key is getting updated.
Like in the following query:
update t1 set d = 72 where b = 2 order by c asc limit 1;

the choice of using index for order by is a key having
(b,c,d) as key_parts.
As d is getting updated, optimizer rejects the index.

This rejection is expected if scanning and updation of
the records are not done in two different passes. But the
current optimizer does take care of this.
Whenever key used for scanning is getting updated,
it first saves the pointers to the required data and
then updates in the second pass. As a result, this limitation
can be lifted.

Solution:
Do not check if used key is being modified when optimizer
chooses index of ordering. Order by always results in two
passes for an update query, whether filesort is used or
not used.