Bug#21658405 ST_UNION() RETURNS AN INVALID POLYGON
The root cause of the bug is that the Boost.Geometry union algorithm does
not take into account certain intersection points of the geometries to be
unioned; as a result, in some cases, the rings produced by the Boost.Geometry
union algorithm contain self-intersection points and do not have the correct
topology.

Fix: add a post-processing step to the Boost.Geometry union algorithm which
takes care of these self intersections; as part of the post-processing step
points may be added on the rings already computed and ring may be split into
smaller rings that do not have self intersection points.

The patch pushed as part of this commit has not yet been pushed into the
Boost.Geometry develop branch and it is still under review. As such, it may
undergo changes, in which case any updates will be migrated to MySQL.