Bug#14695968 NDB_DESC -pb THROWS A SEGMENTATION FAULT
When ndb_desc -pb command was executed on a table containing a
tinyblob column , a segmentation fault occurred.
Blob types might have additional blob tables in addition to the
column.
When partition information for a tinyblob column is printed,
a segmentation fault occurs since the pointer to blob table is
NULL for tinyblob, which is dereferenced in the function that
prints partition info.

Modified ndb_desc so that it checks for NULL condition before
the function that prints partition info is executed. Hence,
the partition info printing function is executed only for columns
that have blob tables.