Bug#21663612 AUTO INCREMENT COLUMN VALUES ARE NOT VISIBLE IN A TABLE CONTAINING SPATIAL INDEX
Issue:
When the primary key columns are the only target columns to return in a select
statement, and the storage engine storing the table has the needed capability,
MySQL optimizer will do a full index scan even if the index is a spatial
index. However, Innodb spatial index doesn't support full index scan, and it returns
HA_ERR_KEY_NOT_FOUND in this case so no result was returned.

Fix:
Don't consider spatial indexes when choosing a target to do full index scan.