Bug#21614368 CRASH IN GEOMETRY::CONSTRUCT WITH BLOB INPUT
Issue: The ST_ASWKB() function was given a Field_blob object to read GEOMETRY
data from, but the row buffer doesn't contain valid row data, for blob fields
that means the blob buffer pointer&length stored in the row buffer is invalid,
so when ST_ASWKB() tries to access the GEOMETRY data at this invalid address,
segment fault occurs.

Fix: GEOMETRY data is never convertible to date[time] data or any other type of
data, we know this for sure, so don't try to detect whether it can be
convertible in the first place so above issue is avoided.

Doing so works around the above issue but
doesn't directly solve it --- to solve it we would need to maintain a state in
TABLE class indicating whether the row buffer has valid row data or not, but
that's beyond this bug fix I would think.

Also, make the field type of ST_ASWKB() as GEOMETRY. Although the returned blob
is actually a WKB byte string rather than a GEOMETRY byte string, the
difference is only relevant to GIS functions, which do not rely on the field
type anyway --- when we access any blob data we check its content&structure to
make sure it is valid GEOMETRY data, so GIS functions won't be fooled by the
field type. To the rest of MySQL where the field type really is concerned,
WKB data is also geometry data which can't be converted to any other type of
data either, and is not applicable anywhere the GEOMETRY data isn't applicable.

To distinguish functions producing WKB data from those producing GEOMETRY data,
add a SP_WKB_FUNC function type and let ST_ASWKB() be of such a function type.