Bug#21619634 ACCESS OF DEALLOCATED ITEM_OUTER_REF CAUSES CRASH: PREPARED STMT
An Item_outer_ref is substituted for the original Item_field as part of
binding.  This is registered for rollback with a view to separate prepared
statements via the usual mechanism: THD::change_item_tree, in
Item_field::fix_outer_field, ca line 5548:

  if (!(rf= new Item_outer_ref(context, this))) ..
  thd->change_item_tree(reference, rf);

A semi-join tranformation produces a permanent node, Item_func_eq ca line 2118
in sql_resolver.cc, which takes as its left item the transient Item_outer_ref:

  new Item_func_eq(in_subq_pred->left_expr->element_index(i),
                   subq_select->ref_pointer_array[i]);

However, this is not registered for rollback, and when we come to execution,
this pointer points to garbage, hence the crash.

The patch registers this location (in Item_func_eq) as the one which needs the
rollback, instead of the original location, which has now been made irrelevant
by the transformation. This lets re-binding at execute time proceed without
seeing the transient object, and all is well.

A new method,

  replace_rollback_place_for_value

is introduced for this purpose. A related method, change_item_tree_place has
been renamed to

   replace_rollback_place_for_ref

to show the relationship between them.

We also removed two redundant arguments for nocheck_register_item_tree_change
and added new constructor to Item_change_record.

The repro test has been added to subquery_sj.inc.