Bug#21205577 ASSERTION FAILED: (INT)IDX >= 0 && IDX < PARENT_JOIN->TABLES
We have
SELECT /* A */
  (SELECT 1 IN (SELECT 1 FROM x1))
  IN  /* B */
  (
    SELECT 1 FROM t2
    WHERE a IN (SELECT 4 FROM x2)
  )
FROM DUAL;
x2 gets semi-join merged, so the inner subquery has two tables t2
and x2. The IN marked with (B) is changed to EXISTS, so
1=(SELECT 1 IN (SELECT 1 FROM x1))
is attached to t2 or x2; it is informed of that fact by
inform_item_in_cond_of_tab(number of t2 or x2, >0).
Then 1 IN (SELECT 1 FROM x1) needs to decide if it will use
subquery materialization, for that it accesses its outer
SELECT, to find the table of number in_cond_of_tab:
this should be the SELECT of t2/x2, but outer_select()
is the one marked (A) which is table-less. The table can't be found,
an assert fails.

Fix: in_cond_of_tab was set wrongly. See comment in item_subselect.cc.
Don't set it if JOIN_TAB is from an inner query block.