Bug#21628840: CRASH/MEMORY CORRUPTION ADDING INDEXES TO               VIRTUAL COLUMN (II)
When an index is added on a virtual generated column, the
algorithm building the index would read the values needed to
generate the value(s). If the depended-upon column was
stored, generated or not, only the values for the column
would be read. But the callback function that calculates
virtual generated column values would try to recalculate
even stored generated columns, despite the fact that the
algorithm didn't read the values.

Fixed by not regenerating stored values in the callback.