Bug#21185883 MISSING LIBSTLPORT.SO LINK IN SOLARISSTUDIO 12.4 VERSION
This is a (partial) backport of
  Bug#16555106 FIX BROKEN BUILD WITH SOLARIS/GCC 64BIT MODE
  Bug#17826757 CMAKE SUPPORT FOR GCC ON SOLARIS
  Bug#17954277 BUILD FAILS ON SOLARIS IF NO STL_LIBRARY_NAME FOUND
  Bug#19010286 BUILD WITH SUN C++ 5.13 SUNOS_SPARC BETA

This adds support for building with Solaris Studio 5.13

When building with stlport4 we install the runtime libraries,
since these are part of the Sun Studio, and not part of Solaris by defualt.
When building with stdcxx4, we get dependencies on libstdcxx4.so
which is part of Solaris by default.

The only new code is:
 - an extra -R'\$ORIGIN/..' in CMAKE_SHARED_LIBRARY_C_FLAGS
   so that plugins are able to find the stlport C++ library
 - new compile flags when cmake is invoked with -DSUNPRO_CXX_LIBRARY=stdcxx4

Additionally:
 - disable building of unittest/gunit/stdcxx, it was exeperimental,
   and un-used anyways.
 - #include <ios> in validate_password.cc