Bug #19893908	: HANDLE_FATAL_SIGNAL (SIG=6) IN FILESORT | SQL/FILESORT.CC:415
Problem:
While doing filesort for the update query, optimizer tries to access a
variable having a stale pointer and exits.

Solution:
In table::init() call, reginfo->join_tab should be initialized to NULL. Else
there is a chance that it will have a stale pointer stored as part of the
previous query execution.

Fixed in 5.7 as part of WL#6042 "split JOIN_TAB into optimizer and executor part"

The problem does not exist in 5.5 as the plans and the code path taken has
changed in 5.6. The table->reginfo.join_tab gets set to NULL in
JOIN_TAB::cleanup. But this is done only when JOIN_TAB is not NULL. While in
5.5, this does not get set to NULL before JOIN_TAB::cleanup, in 5.6 it does
happen. Hence the crash later. The call to
subselect_single_select_engine::prepare() before or after JOIN::cleanup makes
the difference.